itest/test: move test hold persistence into its own file
